Moves RenderCache to renderers directory
