+++ lib/windshaft/renderers/render_cache.js
@@ -1,9 +1,9 @@
 // caches Mapnik render objects, purging after 60 seconds of inactivity:
 var   _                     = require('underscore')
-    , CacheEntry            = require('./cache/cache_entry')
+    , CacheEntry            = require('../cache/cache_entry')
     , Step                  = require('step')
     , tilelive              = require('tilelive')
-    , TorqueRendererFactory = require('./renderers/torque').factory;
+    , TorqueRendererFactory = require('./torque').factory;
 
 require('tilelive-mapnik').registerProtocols(tilelive);
