+++ lib/windshaft/server.js
@@ -121,7 +121,7 @@ module.exports = function(opts){
     var map_store  = new MapStore(map_store_opts);
 
     // initialize render cache
-    var RenderCache = require('./render_cache');
+    var RenderCache = require('./renderers/render_cache');
     var renderCacheOpts = _.defaults(opts.renderCache || {}, {
         ttl: 60000, // 60 seconds TTL by default
         statsInterval: 60000 // reports stats every milliseconds defined here
