+++ test/unit/render_cache.test.js
@@ -3,7 +3,7 @@ var   _             = require('underscore')
     , th            = require('../support/test_helper.js')
     , assert        = require('assert')
     , grainstore    = require('grainstore')
-    , RenderCache   = require('../../lib/windshaft/render_cache.js')
+    , RenderCache   = require('../../lib/windshaft/renderers/render_cache')
     , redis         = require('redis')
     , Step          = require('step')
     , serverOptions = require('../support/server_options')
